Fix postcss plugin
