+++ Green-Wxrp-Shop/postcss.config.mjs
@@ -1,8 +1,5 @@
 const config = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
+  plugins: ["@tailwindcss/postcss"],
 };
 
 export default config;
